added confirmation before delete

+++ src/app/get-bug/get-bug.component.ts
@@ -21,10 +21,11 @@ export class GetBugComponent implements OnInit {
   bugArray: Bug[]=[];
   toggleEllipses:false;
 
-
-
-
   deleteBug(bugId){
+    let ask=confirm("Do you want to delete bug this bug ?");
+    if (!ask){
+      return;
+    }
     this.bugService.deleteBug(bugId).subscribe(response => {
       this.bugList = response;
       console.log(response);
